add test main
