+++ tests/data_test_main/user_inputs/pvcompare_inputs/building_parameters.csv
